categoria en criterios, actualziacion puntajes en todo el app, retirar estrellas, retirar imagenes estrellas y nombres de las imagenes del app, ocultar mantenimiento estrella y reconocimientos, reestablecimientos de las rutas para imagenes de reconocimiento a insignia podria afectar a los reportes q muestran imagenes

+++ back-end/sres/sres.be/ConvocatoriaCriterioPuntajeBE.cs
@@ -12,7 +12,7 @@ namespace sres.be
         public int ID_CRITERIO { get; set; }
         public int ID_DETALLE { get; set; }
         public string DESCRIPCION { get; set; }
-        public int PUNTAJE { get; set; }
+        public decimal PUNTAJE { get; set; }
         public decimal EMISIONES_REDUCIDAS { get; set; }
         public decimal ENERGIA { get; set; }
         public decimal COMBUSTIBLE { get; set; }
