+++ back-end/sres/sres.be/ConvocatoriaEtapaInscripcionBE.cs
@@ -14,7 +14,7 @@ namespace sres.be
         public string REALIZADO { get; set; }
         public int ID_TIPO_EVALUACION { get; set; }
         public string OBSERVACION { get; set; }
-        public int PUNTAJE { get; set; }
+        public decimal PUNTAJE { get; set; }
         public decimal EMISIONES_REDUCIDAS { get; set; }
         public decimal ENERGIA { get; set; }
         public decimal COMBUSTIBLE { get; set; }
