+++ back-end/sres/sres.be/ConvocatoriaInsigniaBE.cs
@@ -12,7 +12,7 @@ namespace sres.be
         public int ID_INSIGNIA { get; set; }
         public string NOMBRE_INSIGNIA { get; set; }
         public string ARCHIVO_BASE { get; set; }
-        public int PUNTAJE_MIN { get; set; }
+        public decimal PUNTAJE_MIN { get; set; }
         public string FLAG_ESTADO { get; set; }
     }
 }
