+++ back-end/sres/sres.be/InsigniaBE.cs
@@ -10,7 +10,7 @@ namespace sres.be
     {
         public int ID_INSIGNIA { get; set; }
         public string NOMBRE { get; set; }
-        public int PUNTAJE_MIN { get; set; }
+        public decimal PUNTAJE_MIN { get; set; }
         public string ARCHIVO_BASE { get; set; }
         public string ARCHIVO_CIFRADO { get; set; }
         public byte[] ARCHIVO_CONTENIDO { get; set; }
