+++ back-end/sres/sres.be/PuntajeBE.cs
@@ -11,7 +11,7 @@ namespace sres.be
         public int ID_CRITERIO { get; set; }
         public int ID_DETALLE { get; set; }
         public string DESCRIPCION { get; set; }
-        public int PUNTAJE { get; set; }
+        public decimal PUNTAJE { get; set; }
         public string CRITERIO { get; set; }
         public string FLAG_ESTADO { get; set; }
     }
