+++ back-end/sres/sres.be/ReconocimientoBE.cs
@@ -15,7 +15,7 @@ namespace sres.be
         public InsigniaBE INSIGNIA { get; set; }
         public EstrellaBE ESTRELLA_E { get; set; }
         public List<ReconocimientoMedidaBE> LISTA_REC_MEDMIT { get; set; }
-        public int PUNTAJE { get; set; }
+        public decimal PUNTAJE { get; set; }
         public string FLAG_CATEGORIA { get; set; }
         public int? ID_ESTRELLA { get; set; }
         public decimal EMISIONES { get; set; }
